Split find unit tests into header and source
Splits into header and source
Replaces std::size_t with int for Vc datapar compatibility
