+++ libs/core/algorithms/tests/unit/algorithms/find.cpp
@@ -1,3 +1,4 @@
+//  Copyright (c) 2021 Srinivas Yadav
 //  Copyright (c) 2014 Grant Mercer
 //
 //  SPDX-License-Identifier: BSL-1.0
@@ -5,86 +6,14 @@
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
 #include <hpx/local/init.hpp>
-#include <hpx/modules/testing.hpp>
-#include <hpx/parallel/algorithms/find.hpp>
 
-#include <cstddef>
 #include <iostream>
-#include <iterator>
-#include <numeric>
-#include <random>
 #include <string>
 #include <vector>
 
-#include "test_utils.hpp"
+#include "find_tests.hpp"
 
 ////////////////////////////////////////////////////////////////////////////
-unsigned int seed = std::random_device{}();
-std::mt19937 gen(seed);
-std::uniform_int_distribution<> dis(2, 101);
-
-template <typename IteratorTag>
-void test_find(IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c(10007);
-    //fill vector with random values about 1
-    std::fill(std::begin(c), std::end(c), dis(gen));
-    c.at(c.size() / 2) = 1;
-
-    iterator index = hpx::find(
-        iterator(std::begin(c)), iterator(std::end(c)), std::size_t(1));
-
-    base_iterator test_index = std::begin(c) + c.size() / 2;
-
-    HPX_TEST(index == iterator(test_index));
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_find(ExPolicy&& policy, IteratorTag)
-{
-    static_assert(hpx::is_execution_policy<ExPolicy>::value,
-        "hpx::is_execution_policy<ExPolicy>::value");
-
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c(10007);
-    //fill vector with random values about 1
-    std::fill(std::begin(c), std::end(c), dis(gen));
-    c.at(c.size() / 2) = 1;
-
-    iterator index = hpx::find(
-        policy, iterator(std::begin(c)), iterator(std::end(c)), std::size_t(1));
-
-    base_iterator test_index = std::begin(c) + c.size() / 2;
-
-    HPX_TEST(index == iterator(test_index));
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_find_async(ExPolicy&& p, IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c(10007);
-    //fill vector with random values above 1
-    std::fill(std::begin(c), std::end(c), dis(gen));
-    c.at(c.size() / 2) = 1;
-
-    hpx::future<iterator> f = hpx::find(
-        p, iterator(std::begin(c)), iterator(std::end(c)), std::size_t(1));
-    f.wait();
-
-    //create iterator at position of value to be found
-    base_iterator test_index = std::begin(c) + c.size() / 2;
-
-    HPX_TEST(f.get() == iterator(test_index));
-}
-
 template <typename IteratorTag>
 void test_find()
 {
@@ -106,112 +35,7 @@ void find_test()
     test_find<std::forward_iterator_tag>();
 }
 
-///////////////////////////////////////////////////////////////////////////////
-template <typename IteratorTag>
-void test_find_exception(IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::decorated_iterator<base_iterator, IteratorTag>
-        decorated_iterator;
-    std::vector<std::size_t> c(10007);
-    std::iota(std::begin(c), std::end(c), gen() + 1);
-    c[c.size() / 2] = 0;
-
-    bool caught_exception = false;
-    try
-    {
-        hpx::find(decorated_iterator(std::begin(c),
-                      []() { throw std::runtime_error("test"); }),
-            decorated_iterator(std::end(c)), std::size_t(0));
-        HPX_TEST(false);
-    }
-    catch (hpx::exception_list const& e)
-    {
-        caught_exception = true;
-        test::test_num_exceptions<hpx::execution::sequenced_policy,
-            IteratorTag>::call(hpx::execution::seq, e);
-    }
-    catch (...)
-    {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_exception);
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_find_exception(ExPolicy&& policy, IteratorTag)
-{
-    static_assert(hpx::is_execution_policy<ExPolicy>::value,
-        "hpx::is_execution_policy<ExPolicy>::value");
-
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::decorated_iterator<base_iterator, IteratorTag>
-        decorated_iterator;
-    std::vector<std::size_t> c(10007);
-    std::iota(std::begin(c), std::end(c), gen() + 1);
-    c[c.size() / 2] = 0;
-
-    bool caught_exception = false;
-    try
-    {
-        hpx::find(policy,
-            decorated_iterator(
-                std::begin(c), []() { throw std::runtime_error("test"); }),
-            decorated_iterator(std::end(c)), std::size_t(0));
-        HPX_TEST(false);
-    }
-    catch (hpx::exception_list const& e)
-    {
-        caught_exception = true;
-        test::test_num_exceptions<ExPolicy, IteratorTag>::call(policy, e);
-    }
-    catch (...)
-    {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_exception);
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_find_exception_async(ExPolicy&& p, IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::decorated_iterator<base_iterator, IteratorTag>
-        decorated_iterator;
-
-    std::vector<std::size_t> c(10007);
-    std::iota(std::begin(c), std::end(c), gen() + 1);
-    c[c.size() / 2] = 0;
-
-    bool caught_exception = false;
-    bool returned_from_algorithm = false;
-    try
-    {
-        hpx::future<decorated_iterator> f = hpx::find(p,
-            decorated_iterator(
-                std::begin(c), []() { throw std::runtime_error("test"); }),
-            decorated_iterator(std::end(c)), std::size_t(0));
-        returned_from_algorithm = true;
-        f.get();
-
-        HPX_TEST(false);
-    }
-    catch (hpx::exception_list const& e)
-    {
-        caught_exception = true;
-        test::test_num_exceptions<ExPolicy, IteratorTag>::call(p, e);
-    }
-    catch (...)
-    {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_exception);
-    HPX_TEST(returned_from_algorithm);
-}
-
+////////////////////////////////////////////////////////////////////////////
 template <typename IteratorTag>
 void test_find_exception()
 {
@@ -235,77 +59,7 @@ void find_exception_test()
     test_find_exception<std::forward_iterator_tag>();
 }
 
-//////////////////////////////////////////////////////////////////////////////
-template <typename ExPolicy, typename IteratorTag>
-void test_find_bad_alloc(ExPolicy&& policy, IteratorTag)
-{
-    static_assert(hpx::is_execution_policy<ExPolicy>::value,
-        "hpx::is_execution_policy<ExPolicy>::value");
-
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::decorated_iterator<base_iterator, IteratorTag>
-        decorated_iterator;
-
-    std::vector<std::size_t> c(100007);
-    std::iota(std::begin(c), std::end(c), gen() + 1);
-    c[c.size() / 2] = 0;
-
-    bool caught_bad_alloc = false;
-    try
-    {
-        hpx::find(policy,
-            decorated_iterator(std::begin(c), []() { throw std::bad_alloc(); }),
-            decorated_iterator(std::end(c)), std::size_t(0));
-        HPX_TEST(false);
-    }
-    catch (std::bad_alloc const&)
-    {
-        caught_bad_alloc = true;
-    }
-    catch (...)
-    {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_bad_alloc);
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_find_bad_alloc_async(ExPolicy&& p, IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::decorated_iterator<base_iterator, IteratorTag>
-        decorated_iterator;
-
-    std::vector<std::size_t> c(10007);
-    std::iota(std::begin(c), std::end(c), gen() + 1);
-    c[c.size() / 2] = 0;
-
-    bool caught_bad_alloc = false;
-    bool returned_from_algorithm = false;
-    try
-    {
-        hpx::future<decorated_iterator> f = hpx::find(p,
-            decorated_iterator(std::begin(c), []() { throw std::bad_alloc(); }),
-            decorated_iterator(std::end(c)), std::size_t(0));
-        returned_from_algorithm = true;
-        f.get();
-
-        HPX_TEST(false);
-    }
-    catch (std::bad_alloc const&)
-    {
-        caught_bad_alloc = true;
-    }
-    catch (...)
-    {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_bad_alloc);
-    HPX_TEST(returned_from_algorithm);
-}
-
+////////////////////////////////////////////////////////////////////////////
 template <typename IteratorTag>
 void test_find_bad_alloc()
 {
