+++ libs/core/algorithms/tests/unit/algorithms/findend.cpp
@@ -1,105 +1,17 @@
-//  Copyright (c) 2014 Grant Mercer
+//  Copyright (c) 2021 Srinivas Yadav
+//  copyright (c) 2014 Grant Mercer
 //
 //  SPDX-License-Identifier: BSL-1.0
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
 #include <hpx/local/init.hpp>
-#include <hpx/modules/testing.hpp>
-#include <hpx/parallel/algorithms/find.hpp>
 
-#include <cstddef>
-#include <functional>
 #include <iostream>
-#include <iterator>
-#include <numeric>
-#include <random>
 #include <string>
 #include <vector>
 
-#include "test_utils.hpp"
-
-////////////////////////////////////////////////////////////////////////////
-unsigned int seed = std::random_device{}();
-std::mt19937 gen(seed);
-std::uniform_int_distribution<> dis(3, 102);
-std::uniform_int_distribution<> dist(7, 106);
-
-template <typename IteratorTag>
-void test_find_end1(IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c(10007);
-    // fill vector with random values above 2
-    std::fill(std::begin(c), std::end(c), dis(gen));
-    // create subsequence in middle of vector
-    c[c.size() / 2] = 1;
-    c[c.size() / 2 + 1] = 2;
-
-    std::size_t h[] = {1, 2};
-
-    iterator index = hpx::find_end(iterator(std::begin(c)),
-        iterator(std::end(c)), std::begin(h), std::end(h));
-
-    iterator test_index = std::find_end(iterator(std::begin(c)),
-        iterator(std::end(c)), std::begin(h), std::end(h));
-
-    HPX_TEST(index == test_index);
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_find_end1(ExPolicy&& policy, IteratorTag)
-{
-    static_assert(hpx::is_execution_policy<ExPolicy>::value,
-        "hpx::is_execution_policy<ExPolicy>::value");
-
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c(10007);
-    // fill vector with random values above 2
-    std::fill(std::begin(c), std::end(c), dis(gen));
-    // create subsequence in middle of vector
-    c[c.size() / 2] = 1;
-    c[c.size() / 2 + 1] = 2;
-
-    std::size_t h[] = {1, 2};
-
-    iterator index = hpx::find_end(policy, iterator(std::begin(c)),
-        iterator(std::end(c)), std::begin(h), std::end(h));
-
-    iterator test_index = std::find_end(iterator(std::begin(c)),
-        iterator(std::end(c)), std::begin(h), std::end(h));
-
-    HPX_TEST(index == test_index);
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_find_end1_async(ExPolicy&& p, IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c(10007);
-    // fill vector with random values above 2
-    std::fill(std::begin(c), std::end(c), dis(gen));
-    // create subsequence in middle of vector
-    c[c.size() / 2] = 1;
-    c[c.size() / 2 + 1] = 2;
-
-    std::size_t h[] = {1, 2};
-
-    hpx::future<iterator> f = hpx::find_end(p, iterator(std::begin(c)),
-        iterator(std::end(c)), std::begin(h), std::end(h));
-    f.wait();
-
-    iterator test_index = std::find_end(iterator(std::begin(c)),
-        iterator(std::end(c)), std::begin(h), std::end(h));
-
-    HPX_TEST(f.get() == test_index);
-}
+#include "findend_tests.hpp"
 
 template <typename IteratorTag>
 void test_find_end1()
@@ -120,89 +32,6 @@ void find_end_test1()
 {
     test_find_end1<std::random_access_iterator_tag>();
     test_find_end1<std::forward_iterator_tag>();
-}
-
-template <typename IteratorTag>
-void test_find_end2(IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c(10007);
-    // fill vector with random values about 2
-    std::fill(std::begin(c), std::end(c), dis(gen));
-    // create subsequence at start and end
-    c[0] = 1;
-    c[1] = 2;
-    c[c.size() - 1] = 2;
-    c[c.size() - 2] = 1;
-
-    std::size_t h[] = {1, 2};
-
-    iterator index = hpx::find_end(iterator(std::begin(c)),
-        iterator(std::end(c)), std::begin(h), std::end(h));
-
-    iterator test_index = std::find_end(iterator(std::begin(c)),
-        iterator(std::end(c)), std::begin(h), std::end(h));
-
-    HPX_TEST(index == test_index);
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_find_end2(ExPolicy&& policy, IteratorTag)
-{
-    static_assert(hpx::is_execution_policy<ExPolicy>::value,
-        "hpx::is_execution_policy<ExPolicy>::value");
-
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c(10007);
-    // fill vector with random values about 2
-    std::fill(std::begin(c), std::end(c), dis(gen));
-    // create subsequence at start and end
-    c[0] = 1;
-    c[1] = 2;
-    c[c.size() - 1] = 2;
-    c[c.size() - 2] = 1;
-
-    std::size_t h[] = {1, 2};
-
-    iterator index = hpx::find_end(policy, iterator(std::begin(c)),
-        iterator(std::end(c)), std::begin(h), std::end(h));
-
-    iterator test_index = std::find_end(iterator(std::begin(c)),
-        iterator(std::end(c)), std::begin(h), std::end(h));
-
-    HPX_TEST(index == test_index);
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_find_end2_async(ExPolicy&& p, IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    // fill vector with random values above 2
-    std::vector<std::size_t> c(10007);
-    std::fill(std::begin(c), std::end(c), dis(gen));
-
-    // create subsequence at start and end
-    c[0] = 1;
-    c[1] = 2;
-    c[c.size() - 1] = 2;
-    c[c.size() - 2] = 1;
-
-    std::size_t h[] = {1, 2};
-
-    hpx::future<iterator> f = hpx::find_end(p, iterator(std::begin(c)),
-        iterator(std::end(c)), std::begin(h), std::end(h));
-    f.wait();
-
-    iterator test_index = std::find_end(iterator(std::begin(c)),
-        iterator(std::end(c)), std::begin(h), std::end(h));
-
-    HPX_TEST(f.get() == test_index);
 }
 
 template <typename IteratorTag>
@@ -227,90 +56,6 @@ void find_end_test2()
 }
 
 template <typename IteratorTag>
-void test_find_end3(IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    // fill vector with random values above 2
-    std::vector<std::size_t> c(10007);
-    std::fill(std::begin(c), std::end(c), dis(gen));
-
-    // create subsequence large enough to always be split into multiple partitions
-    std::iota(std::begin(c), std::begin(c) + c.size() / 16 + 1, 1);
-    std::size_t sub_size = c.size() / 16 + 1;
-
-    std::vector<std::size_t> h(sub_size);
-    std::iota(std::begin(h), std::end(h), 1);
-
-    iterator index = hpx::find_end(iterator(std::begin(c)),
-        iterator(std::end(c)), std::begin(h), std::end(h));
-
-    iterator test_index = std::find_end(iterator(std::begin(c)),
-        iterator(std::end(c)), std::begin(h), std::end(h));
-
-    HPX_TEST(index == test_index);
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_find_end3(ExPolicy&& policy, IteratorTag)
-{
-    static_assert(hpx::is_execution_policy<ExPolicy>::value,
-        "hpx::is_execution_policy<ExPolicy>::value");
-
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    // fill vector with random values above 2
-    std::vector<std::size_t> c(10007);
-    std::fill(std::begin(c), std::end(c), dis(gen));
-
-    // create subsequence large enough to always be split into multiple partitions
-    std::iota(std::begin(c), std::begin(c) + c.size() / 16 + 1, 1);
-    std::size_t sub_size = c.size() / 16 + 1;
-
-    std::vector<std::size_t> h(sub_size);
-    std::iota(std::begin(h), std::end(h), 1);
-
-    iterator index = hpx::find_end(policy, iterator(std::begin(c)),
-        iterator(std::end(c)), std::begin(h), std::end(h));
-
-    iterator test_index = std::find_end(iterator(std::begin(c)),
-        iterator(std::end(c)), std::begin(h), std::end(h));
-
-    HPX_TEST(index == test_index);
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_find_end3_async(ExPolicy&& p, IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    // fill vector with random values above 6
-    std::vector<std::size_t> c(10007);
-    std::fill(std::begin(c), std::end(c), dist(gen));
-
-    // create subsequence large enough to always be split into multiple partitions
-    std::iota(std::begin(c), std::begin(c) + c.size() / 16 + 1, 1);
-    std::size_t sub_size = c.size() / 16 + 1;
-
-    std::vector<std::size_t> h(sub_size);
-    std::iota(std::begin(h), std::end(h), 1);
-
-    // create only two partitions, splitting the desired sub sequence into
-    // separate partitions.
-    hpx::future<iterator> f = hpx::find_end(p, iterator(std::begin(c)),
-        iterator(std::end(c)), std::begin(h), std::end(h));
-    f.wait();
-
-    iterator test_index = std::find_end(iterator(std::begin(c)),
-        iterator(std::end(c)), std::begin(h), std::end(h));
-
-    HPX_TEST(f.get() == test_index);
-}
-
-template <typename IteratorTag>
 void test_find_end3()
 {
     using namespace hpx::execution;
@@ -329,90 +74,6 @@ void find_end_test3()
 {
     test_find_end3<std::random_access_iterator_tag>();
     test_find_end3<std::forward_iterator_tag>();
-}
-
-template <typename IteratorTag>
-void test_find_end4(IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c(10007);
-    // fill vector with random values above 2
-    std::fill(std::begin(c), std::end(c), dis(gen));
-
-    // create subsequence in middle of vector
-    c[c.size() / 2] = 1;
-    c[c.size() / 2 + 1] = 2;
-
-    std::size_t h[] = {1, 2};
-
-    iterator index = hpx::find_end(iterator(std::begin(c)),
-        iterator(std::end(c)), std::begin(h), std::end(h),
-        [](std::size_t v1, std::size_t v2) { return !(v1 != v2); });
-
-    iterator test_index = std::find_end(iterator(std::begin(c)),
-        iterator(std::end(c)), std::begin(h), std::end(h),
-        [](std::size_t v1, std::size_t v2) { return !(v1 != v2); });
-
-    HPX_TEST(index == test_index);
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_find_end4(ExPolicy&& policy, IteratorTag)
-{
-    static_assert(hpx::is_execution_policy<ExPolicy>::value,
-        "hpx::is_execution_policy<ExPolicy>::value");
-
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    std::vector<std::size_t> c(10007);
-    // fill vector with random values above 2
-    std::fill(std::begin(c), std::end(c), dis(gen));
-    // create subsequence in middle of vector
-    c[c.size() / 2] = 1;
-    c[c.size() / 2 + 1] = 2;
-
-    std::size_t h[] = {1, 2};
-
-    iterator index = hpx::find_end(policy, iterator(std::begin(c)),
-        iterator(std::end(c)), std::begin(h), std::end(h),
-        [](std::size_t v1, std::size_t v2) { return !(v1 != v2); });
-
-    iterator test_index = std::find_end(iterator(std::begin(c)),
-        iterator(std::end(c)), std::begin(h), std::end(h),
-        [](std::size_t v1, std::size_t v2) { return !(v1 != v2); });
-
-    HPX_TEST(index == test_index);
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_find_end4_async(ExPolicy&& p, IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
-
-    // fill vector with random values above 2
-    std::vector<std::size_t> c(10007);
-    std::fill(std::begin(c), std::end(c), dis(gen));
-
-    // create subsequence in middle of vector
-    c[c.size() / 2] = 1;
-    c[c.size() / 2 + 1] = 2;
-
-    std::size_t h[] = {1, 2};
-
-    hpx::future<iterator> f = hpx::find_end(p, iterator(std::begin(c)),
-        iterator(std::end(c)), std::begin(h), std::end(h),
-        [](std::size_t v1, std::size_t v2) { return !(v1 != v2); });
-    f.wait();
-
-    iterator test_index = std::find_end(iterator(std::begin(c)),
-        iterator(std::end(c)), std::begin(h), std::end(h),
-        [](std::size_t v1, std::size_t v2) { return !(v1 != v2); });
-
-    HPX_TEST(f.get() == test_index);
 }
 
 template <typename IteratorTag>
@@ -436,132 +97,6 @@ void find_end_test4()
     test_find_end4<std::forward_iterator_tag>();
 }
 
-///////////////////////////////////////////////////////////////////////////////
-template <typename ExPolicy, typename IteratorTag>
-void test_find_end_exception(ExPolicy&& policy, IteratorTag)
-{
-    static_assert(hpx::is_execution_policy<ExPolicy>::value,
-        "hpx::is_execution_policy<ExPolicy>::value");
-
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::decorated_iterator<base_iterator, IteratorTag>
-        decorated_iterator;
-    std::vector<std::size_t> c(10007);
-    std::iota(std::begin(c), std::end(c), gen() + 1);
-    c[c.size() / 2] = 1;
-    c[c.size() / 2 + 1] = 2;
-
-    std::vector<std::size_t> h;
-    h.push_back(1);
-    h.push_back(2);
-
-    bool caught_exception = false;
-    try
-    {
-        hpx::find_end(policy,
-            decorated_iterator(
-                std::begin(c), []() { throw std::runtime_error("test"); }),
-            decorated_iterator(
-                std::end(c), []() { throw std::runtime_error("test"); }),
-            std::begin(h), std::end(h));
-        HPX_TEST(false);
-    }
-    catch (hpx::exception_list const& e)
-    {
-        caught_exception = true;
-        test::test_num_exceptions<ExPolicy, IteratorTag>::call(policy, e);
-    }
-    catch (...)
-    {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_exception);
-}
-
-template <typename IteratorTag>
-void test_find_end_exception(IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::decorated_iterator<base_iterator, IteratorTag>
-        decorated_iterator;
-
-    std::vector<std::size_t> c(10007);
-    std::iota(std::begin(c), std::end(c), gen() + 1);
-    c[c.size() / 2] = 1;
-    c[c.size() / 2 + 1] = 2;
-
-    std::vector<std::size_t> h;
-    h.push_back(1);
-    h.push_back(2);
-
-    bool caught_exception = false;
-    try
-    {
-        hpx::find_end(decorated_iterator(std::begin(c),
-                          []() { throw std::runtime_error("test"); }),
-            decorated_iterator(
-                std::end(c), []() { throw std::runtime_error("test"); }),
-            std::begin(h), std::end(h));
-        HPX_TEST(false);
-    }
-    catch (hpx::exception_list const& e)
-    {
-        caught_exception = true;
-        test::test_num_exceptions<hpx::execution::sequenced_policy,
-            IteratorTag>::call(hpx::execution::seq, e);
-    }
-    catch (...)
-    {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_exception);
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_find_end_exception_async(ExPolicy&& p, IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::decorated_iterator<base_iterator, IteratorTag>
-        decorated_iterator;
-
-    std::vector<std::size_t> c(10007);
-    std::iota(std::begin(c), std::end(c), gen() + 1);
-    c[c.size() / 2] = 1;
-    c[c.size() / 2 + 1] = 2;
-
-    std::size_t h[] = {1, 2};
-
-    bool caught_exception = false;
-    bool returned_from_algorithm = false;
-    try
-    {
-        hpx::future<decorated_iterator> f = hpx::find_end(p,
-            decorated_iterator(
-                std::begin(c), []() { throw std::runtime_error("test"); }),
-            decorated_iterator(
-                std::end(c), []() { throw std::runtime_error("test"); }),
-            std::begin(h), std::end(h));
-        returned_from_algorithm = true;
-        f.get();
-
-        HPX_TEST(false);
-    }
-    catch (hpx::exception_list const& e)
-    {
-        caught_exception = true;
-        test::test_num_exceptions<ExPolicy, IteratorTag>::call(p, e);
-    }
-    catch (...)
-    {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_exception);
-    HPX_TEST(returned_from_algorithm);
-}
-
 template <typename IteratorTag>
 void test_find_end_exception()
 {
@@ -583,83 +118,6 @@ void find_end_exception_test()
 {
     test_find_end_exception<std::random_access_iterator_tag>();
     test_find_end_exception<std::forward_iterator_tag>();
-}
-
-//////////////////////////////////////////////////////////////////////////////
-template <typename ExPolicy, typename IteratorTag>
-void test_find_end_bad_alloc(ExPolicy&& policy, IteratorTag)
-{
-    static_assert(hpx::is_execution_policy<ExPolicy>::value,
-        "hpx::is_execution_policy<ExPolicy>::value");
-
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::decorated_iterator<base_iterator, IteratorTag>
-        decorated_iterator;
-
-    std::vector<std::size_t> c(100007);
-    std::iota(std::begin(c), std::end(c), gen() + 1);
-    c[c.size() / 2] = 0;
-
-    std::size_t h[] = {1, 2};
-
-    bool caught_bad_alloc = false;
-    try
-    {
-        hpx::find_end(policy,
-            decorated_iterator(std::begin(c), []() { throw std::bad_alloc(); }),
-            decorated_iterator(std::end(c), []() { throw std::bad_alloc(); }),
-            std::begin(h), std::end(h));
-        HPX_TEST(false);
-    }
-    catch (std::bad_alloc const&)
-    {
-        caught_bad_alloc = true;
-    }
-    catch (...)
-    {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_bad_alloc);
-}
-
-template <typename ExPolicy, typename IteratorTag>
-void test_find_end_bad_alloc_async(ExPolicy&& p, IteratorTag)
-{
-    typedef std::vector<std::size_t>::iterator base_iterator;
-    typedef test::decorated_iterator<base_iterator, IteratorTag>
-        decorated_iterator;
-
-    std::vector<std::size_t> c(10007);
-    std::iota(std::begin(c), std::end(c), gen() + 1);
-    c[c.size() / 2] = 0;
-
-    std::size_t h[] = {1, 2};
-
-    bool caught_bad_alloc = false;
-    bool returned_from_algorithm = false;
-    try
-    {
-        hpx::future<decorated_iterator> f = hpx::find_end(p,
-            decorated_iterator(std::begin(c), []() { throw std::bad_alloc(); }),
-            decorated_iterator(std::end(c), []() { throw std::bad_alloc(); }),
-            std::begin(h), std::end(h));
-        returned_from_algorithm = true;
-        f.get();
-
-        HPX_TEST(false);
-    }
-    catch (std::bad_alloc const&)
-    {
-        caught_bad_alloc = true;
-    }
-    catch (...)
-    {
-        HPX_TEST(false);
-    }
-
-    HPX_TEST(caught_bad_alloc);
-    HPX_TEST(returned_from_algorithm);
 }
 
 template <typename IteratorTag>
